+++ src/services/TranslateService.php
@@ -5,7 +5,6 @@ namespace digitalpulsebe\craftmultitranslator\services;
 use Craft;
 use craft\base\Component;
 use craft\base\Element;
-use craft\base\Field;
 use craft\base\FieldInterface;
 use craft\commerce\elements\Product;
 use craft\commerce\elements\Variant;
@@ -13,40 +12,65 @@ use craft\elements\Asset;
 use craft\elements\ContentBlock;
 use craft\elements\Entry;
 use craft\enums\PropagationMethod;
-use craft\fields\Table;
+use craft\errors\ElementNotFoundException;
+use craft\errors\InvalidElementException;
+use craft\errors\UnsupportedSiteException;
 use craft\models\Site;
+use digitalpulsebe\craftmultitranslator\base\FieldSerializer;
 use digitalpulsebe\craftmultitranslator\events\FieldTranslationEvent;
+use digitalpulsebe\craftmultitranslator\events\RegisterSerializersEvent;
 use digitalpulsebe\craftmultitranslator\helpers\ElementHelper;
 use digitalpulsebe\craftmultitranslator\events\ElementTranslationEvent;
+use digitalpulsebe\craftmultitranslator\helpers\SerializerHelper;
 use digitalpulsebe\craftmultitranslator\MultiTranslator;
+use digitalpulsebe\craftmultitranslator\serializers\Matrix as MatrixSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Hyper as HyperSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Text as TextSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\RichText as RichTextSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Table as TableSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\EtherSeo as EtherSeoSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Seomatic as SeomaticSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Vizy as VizySerializer;
+use digitalpulsebe\craftmultitranslator\serializers\ContentBlock as ContentBlockSerializer;
+use digitalpulsebe\craftmultitranslator\serializers\Linkit as LinkitSerializer;
+use Throwable;
+use yii\base\Exception;
+use yii\base\InvalidConfigException;
+use yii\web\ForbiddenHttpException;
 
 class TranslateService extends Component
 {
-    public const EVENT_BEFORE_ELEMENT_TRANSLATION = 'beforeElementTranslation';
-    public const EVENT_AFTER_ELEMENT_TRANSLATION = 'afterElementTranslation';
-    public const EVENT_BEFORE_FIELD_TRANSLATION = 'beforeFieldTranslation';
-    public const EVENT_AFTER_FIELD_TRANSLATION = 'afterFieldTranslation';
+    public const string EVENT_BEFORE_ELEMENT_TRANSLATION = 'beforeElementTranslation';
+    public const string EVENT_AFTER_ELEMENT_TRANSLATION = 'afterElementTranslation';
+    public const string EVENT_BEFORE_FIELD_TRANSLATION = 'beforeFieldTranslation';
+    public const string EVENT_AFTER_FIELD_TRANSLATION = 'afterFieldTranslation';
+    public const string EVENT_REGISTER_SERIALIZERS = 'registerTranslationSerializers';
 
-    static array $textFields = [
-        'craft\fields\PlainText',
-        'craft\redactor\Field',
-        'craft\ckeditor\Field',
-        'abmat\tinymce\Field',
-    ];
     static array $matrixFields = [
         'craft\fields\Matrix',
         'benf\neo\Field',
         'verbb\supertable\fields\SuperTableField',
     ];
 
+    protected array $serializers = [];
+
+    public function init(): void
+    {
+        parent::init();
+
+        $this->registerSerializerClasses();
+    }
+
     /**
      * @param Element $source
      * @param Site $sourceSite
      * @param Site $targetSite
-     * @return Element
-     * @throws \Throwable
-     * @throws \craft\errors\ElementNotFoundException
-     * @throws \yii\base\Exception
+     * @param bool $isRootElement
+     * @return Element|null
+     * @throws ElementNotFoundException
+     * @throws Exception
+     * @throws Throwable
+     * @throws InvalidConfigException
      */
     public function translateElement(Element $source, Site $sourceSite, Site $targetSite, bool $isRootElement = true): ?Element
     {
@@ -54,31 +78,13 @@ class TranslateService extends Component
             return null;
         }
 
-        // translate inside of Element, get serialized data
-        $translatedValues = $this->translateElementFields($source, $sourceSite, $targetSite, true);
+        $originalHtml = SerializerHelper::serialize($this->serializeElement($source, $sourceSite, $targetSite));
+        $translatedHtml = $this->translateText($sourceSite->language, $targetSite->language, $originalHtml);
+        $translatedValues = SerializerHelper::unserialize($translatedHtml);
 
         // find or create target (destination)
         $targetElement = $this->findTargetElement($source, $targetSite->id);
-
-        if (isset($translatedValues['title'])) {
-            // title is not a custom field
-            $targetElement->title = $translatedValues['title'];
-
-            if ($this->getProviderSettings()->getResetSlug()) {
-                $targetElement->slug = null;
-            }
-
-            unset($translatedValues['title']);
-        }
-
-        if (isset($translatedValues['alt'])) {
-            // alt is not a custom field
-            $targetElement->alt = $translatedValues['alt'];
-            unset($translatedValues['alt']);
-        }
-
-        // set field values for custom fields
-        $targetElement->setFieldValues($translatedValues);
+        $this->setElementTranslation($source, $targetElement, $translatedValues);
 
         $revisionNotes = 'Translated by Multi Translator from "'
             .$sourceSite->name.'" ('.$sourceSite->getLocale()->getLanguageID().') to "'
@@ -129,6 +135,7 @@ class TranslateService extends Component
                 'propagationMethod' => $source?->section->propagationMethod ?? null,
                 'sourceEntry' => ['id' => $source->id, 'siteId' => $source->siteId, 'draft' => $source->getIsDraft(), 'customFields' => $source->getSerializedFieldValues()],
                 'targetElement' => ['id' => $targetElement->id, 'siteId' => $targetElement->siteId, 'draft' => $targetElement->getIsDraft()],
+                'serialized' => $originalHtml,
                 'translatedValues' => $translatedValues,
             ]);
         }
@@ -146,299 +153,134 @@ class TranslateService extends Component
         return $targetElement;
     }
 
-
     /**
+     * Serialize an element for translation
      * @param Element $source
      * @param Site $sourceSite
      * @param Site $targetSite
      * @return array
+     * @throws InvalidConfigException
      */
-    public function translateElementFields(Element $source, Site $sourceSite, Site $targetSite, bool $isRootElement = false): array
+    public function serializeElement(Element $source, Site $sourceSite, Site $targetSite): array
     {
-        $target = [];
+        $disabledFields = $this->getProviderSettings()->getDisabledFieldHandles();
 
-        $disabledFields = $isRootElement ? $this->getProviderSettings()->getDisabledFieldHandles() : [];
+        $serialized['id'] = $source->id;
 
         if ($source->title && $source->getIsTitleTranslatable() && !in_array('title', $disabledFields)) {
-            $target['title'] = $this->translateText($sourceSite->language, $targetSite->language, $source->title);
+            $serialized['title'] = $source->title;
         }
 
         if ($source instanceof Asset && $source->alt && !in_array($source->getVolume()->altTranslationMethod, ['none', 'custom'])) {
             // assets can have a translatable alt field
-            $target['alt'] = $this->translateText($sourceSite->language, $targetSite->language, $source->alt);
+            $serialized['alt'] = $source->alt;
         }
 
-        $this->processFieldLayoutFields($source, $target, $sourceSite, $targetSite, $disabledFields);
-
-        return $target;
-    }
-
-    protected function processFieldLayoutFields(Element|ContentBlock $source, array &$target, Site $sourceSite, Site $targetSite, array $disabledFields = []): void
-    {
-        foreach ($source->getFieldLayout()->getCustomFields() as $field) {
-            $translatedValue = null;
-            $fieldTranslatable = $field->translationMethod != Field::TRANSLATION_METHOD_NONE;
-            $beforeTranslationEvent = $this->onBeforeFieldTranslation($source, $field, $sourceSite, $targetSite);
-            $processField = boolval($fieldTranslatable); // if translatable
-
-            if (in_array($field->handle, $disabledFields)) {
-                continue;
-            }
-
-            if (!$beforeTranslationEvent->isValid) {
-                // translated value, possibly overridden on event handle
-                $translatedValue = $beforeTranslationEvent->translatedValue;
-            } elseif (in_array(get_class($field), static::$textFields) && $processField) {
-                // normal text fields
-                $translatedValue = $this->translateTextField($source, $field, $sourceSite, $targetSite);
-            } elseif (in_array(get_class($field), static::$matrixFields)) {
-                // dig deeper in Matrix fields
-                $translatedValue = $this->translateMatrixField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == Table::class && $processField) {
-                // loop over table
-                $translatedValue = $this->translateTable($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'lenz\linkfield\fields\LinkField' && $processField) {
-                // translate linkfield custom label
-                $translatedValue = $this->translateLinkField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'craft\fields\ContentBlock' && $processField) {
-                // translate content block
-                $translatedValue = $this->translateContentBlock($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'presseddigital\linkit\fields\LinkitField' && $processField) {
-                // translate Linkit custom text
-                $translatedValue = $this->translateLinkitField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'verbb\hyper\fields\HyperField' && $processField) {
-                // translate hyper linkText
-                $translatedValue = $this->translateHyperField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'ether\seo\fields\SeoField' && $processField) {
-                // translate Ether Seo title and description
-                $translatedValue = $this->translateEtherSeoField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'nystudio107\seomatic\fields\SeoSettings' && $processField) {
-                // translate nystudio107's Seomatic data
-                $translatedValue = $this->translateSeomaticField($source, $field, $sourceSite, $targetSite);
-            } elseif (get_class($field) == 'verbb\vizy\fields\VizyField' && $processField) {
-                // translate nystudio107's Seomatic data
-                $translatedValue = $this->translateVizyField($source, $field, $sourceSite, $targetSite);
-            }
-
-            if (in_array(get_class($field), [
-                'craft\ckeditor\Field',
-                'abmat\tinymce\Field',
-            ])) {
-                // search for interal href links
-                $translatedValue = $this->translateLinks($translatedValue, $sourceSite, $targetSite);
-                $translatedValue = $this->translateNestedEntries($translatedValue, $sourceSite, $targetSite);
-            }
-
-            $translatedValue = $this->onAfterFieldTranslation($source, $field, $sourceSite, $targetSite, $translatedValue);
-
-            if ($translatedValue) {
-                $target[$field->handle] = $translatedValue;
-            } else {
-                $target[$field->handle] = $field->serializeValue($source->getFieldValue($field->handle), $source);
-            }
-        }
-    }
-
-    public function translateTextField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?string
-    {
-        $value = $field->serializeValue($element->getFieldValue($field->handle), $element);
-
-        return $this->translateText($sourceSite->language, $targetSite->language, $value);
-    }
-
-    public function translateTable(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): array
-    {
-        $sourceData = $field->serializeValue($element->getFieldValue($field->handle), $element);
-
-        $targetData = [];
-
-        if (is_array($sourceData)) {
-            $textColumns = [];
-            foreach ($field->columns as $columnName => $columnConfig) {
-                if (in_array($columnConfig['type'], ['singleline', 'multiline', 'heading'])) {
-                    // only process types with text
-                    $textColumns[] = $columnName;
-                }
-            }
-
-            foreach ($sourceData as $sourceRow) {
-                $targetRow = [];
-                foreach ($sourceRow as $columnName => $value) {
-                    if (in_array($columnName, $textColumns)) {
-                        $targetRow[$columnName] = $this->translateText($sourceSite->language, $targetSite->language, $value);
-                    } else {
-                        $targetRow[$columnName] = $value;
-                    }
-                }
-                $targetData[] = $targetRow;
-            }
-        }
-
-        return $targetData;
-    }
-
-    public function translateMatrixField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): array
-    {
-        $query = $element->getFieldValue($field->handle);
-
-        $includeDisabled = $this->getProviderSettings()->getTranslateDisabledMatrixElements();
-        if ($includeDisabled) {
-            $query->status(null);
-        }
-
-        // serialize current value
-        $serialized = $field->serializeValue($query, $element);
-
-        foreach ($query->all() as $matrixElement) {
-            $translatedMatrixValues = $this->translateElementFields($matrixElement, $sourceSite, $targetSite);
-            foreach ($translatedMatrixValues as $matrixFieldHandle => $value) {
-                // only set translated values in matrix array
-                if ($value && isset($serialized[$matrixElement->id])) {
-                    if ($matrixFieldHandle == 'title') {
-                        $serialized[$matrixElement->id][$matrixFieldHandle] = $value;
-                    } else {
-                        $serialized[$matrixElement->id]['fields'][$matrixFieldHandle] = $value;
-                    }
-                }
-            }
-        }
-
-        if (get_class($field) == 'benf\neo\Field' && $field->propagationMethod != PropagationMethod::None) {
-            // special case to avoid neo overwriting blocks in all languages
-            return ['blocks' => $serialized];
-        }
+        $serialized['fields'] = $this->serializeElementFields($source, $sourceSite, $targetSite);
 
         return $serialized;
     }
 
-    public function translateLinkField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
+    /**
+     * Serialize all translatable fields of an element
+     * @param Element|ContentBlock $element
+     * @param Site $sourceSite
+     * @param Site $targetSite
+     * @return array
+     */
+    public function serializeElementFields(Element|ContentBlock $element, Site $sourceSite, Site $targetSite): array
     {
-        $value = $element->getFieldValue($field->handle);
-        if ($value) {
-            try {
-                $array = $value->toArray();
+        return ElementHelper::getElementTranslatableFields($element)
+            ->filter(function ($field)  {
+                return isset($this->serializers[get_class($field)]);
+            })
+            ->map(function ($field) use ($element, $sourceSite, $targetSite) {
+                $serializer = $this->getSerializer($field);
 
-                if (!empty($array['customText'])) {
-                    $array['customText'] = $this->translateText($sourceSite->language, $targetSite->language, $array['customText']);
+                if ($serializer) {
+
+                    $onBeforeFieldTranslationEvent = $this->onBeforeFieldTranslation($element, $field, $sourceSite, $targetSite);
+
+                    if (!$onBeforeFieldTranslationEvent->isValid) {
+                        return null;
+                    }
+
+                    return $serializer->serialize($element, $sourceSite, $targetSite);
                 }
-                $array['type'] = $value->type; // add type, because they forgot to return it with toArray?
-                return $array;
-            } catch (\Throwable $throwable) {
-                // too bad, f*** linkfields
+
                 return null;
-            }
+            })
+            ->all()
+        ;
+    }
+
+    /**
+     * Get the serializer for a field
+     * @param FieldInterface $field
+     * @return FieldSerializer|null
+     */
+    protected function getSerializer(FieldInterface $field): ?FieldSerializer
+    {
+        $fieldClass = get_class($field);
+        $serializerClass = $this->serializers[$fieldClass] ?? null;
+
+        if ($serializerClass) {
+            return new $serializerClass($field);
         }
+
         return null;
     }
 
-    public function translateLinkitField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
+    /**
+     * Set the translation values on an element
+     * @param Element $source
+     * @param Element $target
+     * @param array|null $translatedValues
+     */
+    public function setElementTranslation(Element $source, Element $target, array $translatedValues = null): void
     {
-        $value = $element->getFieldValue($field->handle);
-        if ($field->allowCustomText && $value->customText != null) {
-            $value->customText = $this->translateText($sourceSite->language, $targetSite->language, $value->customText);
+        if (isset($translatedValues['title'])) {
+            $target->title = $translatedValues['title'];
         }
-        return $field->serializeValue($value, $element);
+
+        if (isset($translatedValues['alt'])) {
+            // assets can have a translatable alt field
+            $target->alt = $translatedValues['alt'];
+        }
+
+        if (isset($translatedValues['fields'])) {
+            $target->setFieldValues($this->setElementFieldsTranslations($source, $target, $translatedValues['fields']));
+        }
     }
 
-    public function translateHyperField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
+    /**
+     * Set the translated field values on the fields of an element
+     * @param Element $source
+     * @param Element $target
+     * @param $translatedFieldValues
+     * @return array
+     */
+    public function setElementFieldsTranslations(Element $source, Element $target, $translatedFieldValues): array
     {
-        $values = $element->getSerializedFieldValues()[$field->handle];
-        foreach ($values as $i => $value) {
-            try {
-                if (!empty($value['linkText'])) {
-                    $values[$i]['linkText'] = $this->translateText($sourceSite->language, $targetSite->language, $value['linkText']);
-                }
-                if (!empty($value['linkSiteId']) && $value['linkSiteId'] == $sourceSite->id) {
-                    $values[$i]['linkSiteId'] = $targetSite->id;
-                }
-            } catch (\Throwable $throwable) {
-                // might fail, but we still want to continue
-                MultiTranslator::error([
-                    'message' => 'Error translating Hyper field',
-                    'error' => $throwable->getMessage(),
-                ]);
-            }
-        }
-        return $values;
+        return ElementHelper::getElementTranslatableFields($source)
+            ->filter(function ($field) use ($translatedFieldValues) {
+                return !empty($translatedFieldValues[$field->handle] ?? null);
+            })
+            ->map(function ($field) use ($source, $target, $translatedFieldValues) {
+                $serializer = $this->getSerializer($field);
+                $fieldData = $translatedFieldValues[$field->handle] ?? null;
+                return $serializer?->setFieldData($source, $target, $fieldData);
+            })
+            ->all()
+        ;
     }
 
-    public function translateEtherSeoField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
-    {
-        $value = $element->getFieldValue($field->handle);
-
-        $titles = []; $description = null;
-
-        foreach ($value->titleRaw as $titleValue) {
-            $titles[] = $this->translateText($sourceSite->language, $targetSite->language, $titleValue);
-        }
-
-        if(!empty($value->descriptionRaw)) {
-            $description = $this->translateText($sourceSite->language, $targetSite->language, $value->descriptionRaw);
-        }
-
-        return [
-            'titleRaw' => $titles,
-            'descriptionRaw' => $description
-        ];
-    }
-
-    public function translateSeomaticField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
-    {
-        $serialized = $element->getSerializedFieldValues([$field->handle])[$field->handle];
-
-        $textFields = [
-            'seoTitle', 'seoDescription', 'seoKeywords', 'seoImageDescription',
-            'twitterTitle', 'twitterDescription', 'twitterImageDescription',
-            'ogTitle', 'ogDescription', 'ogImageDescription',
-            ];
-
-        if (isset($serialized['metaBundleSettings'])) {
-            foreach ($textFields as $textField) {
-                $currentValue = $serialized['metaGlobalVars'][$textField] ?? null;
-                if ($currentValue && $serialized['metaBundleSettings'][$textField."Source"] == 'fromCustom') {
-                    $serialized['metaGlobalVars'][$textField] = $this->translateText($sourceSite->language, $targetSite->language, $currentValue);
-                }
-            }
-        }
-
-        return $serialized;
-    }
-
-    public function translateVizyField(Element $element, FieldInterface $field, Site $sourceSite, Site $targetSite): ?array
-    {
-        $nodes = [];
-        foreach ($element->getFieldValue($field->handle)->all() as $vizyNode) {
-            if (get_class($vizyNode) == 'verbb\vizy\nodes\VizyBlock') {
-                $blockElement = $vizyNode->getBlockElement();
-                $translatedMatrixValues = $this->translateElementFields($blockElement, $sourceSite, $targetSite);
-                $node = $vizyNode->serializeValue($blockElement);
-                $node['attrs']['values']['content']['fields'] = $translatedMatrixValues;
-                $nodes[] = $node;
-            } else {
-                // process html content in array
-                $node = $this->translateVizyNode($vizyNode->serializeValue($element), $sourceSite, $targetSite);
-                $nodes[] = $node;
-            }
-        }
-        return $nodes;
-    }
-
-    public function translateVizyNode(array $node, Site $sourceSite, Site $targetSite)
-    {
-        if (isset($node['content']) && is_array($node['content'])) {
-            foreach ($node['content'] as $i => $subNode) {
-                // go deeper
-                $node['content'][$i] = $this->translateVizyNode($subNode, $sourceSite, $targetSite);
-            }
-        }
-
-        if (!empty($node['text'])) {
-            $node['text'] = $this->translateText($sourceSite->language, $targetSite->language, $node['text']);
-        }
-
-        return $node;
-    }
-
+    /**
+     * Find the target element in the target site for a source element
+     * @param Element $source
+     * @param int $targetSiteId
+     * @return Element
+     */
     public function findTargetElement(Element $source, int $targetSiteId): Element
     {
         if ($source instanceof Product) {
@@ -452,6 +294,18 @@ class TranslateService extends Component
         }
     }
 
+    /**
+     * Find the target entry in the target site for a source entry
+     * @param Entry $source
+     * @param int $targetSiteId
+     * @return Entry
+     * @throws ElementNotFoundException
+     * @throws Exception
+     * @throws Throwable
+     * @throws InvalidElementException
+     * @throws UnsupportedSiteException
+     * @throws ForbiddenHttpException
+     */
     public function findTargetEntry(Entry $source, int $targetSiteId): Entry
     {
         $targetEntry = ElementHelper::one(Entry::class, $source->id, $targetSiteId);
@@ -491,6 +345,13 @@ class TranslateService extends Component
         return $targetEntry;
     }
 
+    /**
+     * Translate text (or HTML) using the configured provider
+     * @param string|null $sourceLocale
+     * @param string|null $targetLocale
+     * @param string|null $text
+     * @return string|null
+     */
     public function translateText(string $sourceLocale = null, string $targetLocale = null, string $text = null): ?string
     {
         if ($this->getProviderSettings()->getDetectSourceLanguage()) {
@@ -500,6 +361,10 @@ class TranslateService extends Component
         return $this->getApiService()->translate($sourceLocale, $targetLocale, $text);
     }
 
+    /**
+     * Get the configured translation provider API service
+     * @return ApiService|null
+     */
     public function getApiService(): ?ApiService
     {
         $provider = $this->getProviderSettings()->getTranslationProvider();
@@ -513,86 +378,6 @@ class TranslateService extends Component
         }
 
         return null;
-    }
-
-    public function translateLinks(string $translatedValue = null, Site $sourceSite, Site $targetSite): ?string
-    {
-        if (!$this->getProviderSettings()->getUpdateInternalLinks()) {
-            return $translatedValue;
-        }
-
-        $matches = [];
-        // match pattern like "<a href="{entry:9999@1:url||https://example.com/slug}">link</a>"
-        preg_match_all('/{(entry|asset|variant|product):(\d+)@(\d+):/i', $translatedValue, $matches);
-
-        // should have four arrays: full matches, capture groups 1-3
-        if (count($matches) == 4 && count($matches[0])) {
-            foreach ($matches[0] as $i => $fullMatch) {
-                $type = $matches[1][$i];
-                $entryId = $matches[2][$i];
-                $siteId = $matches[3][$i];
-                $class = null;
-
-                if ($type == 'entry') {
-                    $class = Entry::class;
-                } elseif ($type == 'asset') {
-                    $class = Asset::class;
-                } elseif ($type == 'variant') {
-                    $class = 'craft\commerce\elements\Variant';
-                } elseif ($type == 'product') {
-                    $class = 'craft\commerce\elements\Product';
-                }
-
-                if ($sourceSite->id == $siteId && $class) {
-                    $findTarget = $class::find()->siteId($targetSite->id)->status(null)->id($entryId)->one();
-                    if ($findTarget) {
-                        $targetSiteId = $targetSite->id;
-                        $translatedMatch = '{'.$type.':'.$entryId.'@'.$targetSiteId.':';
-                        $translatedValue = str_replace($fullMatch, $translatedMatch, $translatedValue);
-                    }
-                }
-            }
-        }
-
-        return $translatedValue;
-    }
-
-    public function translateNestedEntries(string $translatedValue = null, Site $sourceSite, Site $targetSite): ?string
-    {
-        if (!$this->getProviderSettings()->getProcessNestedEntries()) {
-            return $translatedValue;
-        }
-
-        $matches = [];
-        preg_match_all('/<craft-entry data-entry-id="(\d+)" ?(data-site-id="\d+")?>/i', $translatedValue, $matches);
-
-        // should have two arrays: full matches, capture group 1-2
-        if (count($matches) == 3 && count($matches[0])) {
-            foreach ($matches[0] as $i => $fullMatch) {
-                $entryId = $matches[1][$i];
-
-                $sourceEntry = Entry::find()->siteId($sourceSite->id)->id($entryId)->one();
-                $targetEntry = $sourceEntry ? $this->translateElement($sourceEntry, $sourceSite, $targetSite, false) : null;
-                if ($targetEntry) {
-                    $translatedValue = str_replace($fullMatch, "<craft-entry data-entry-id=\"$targetEntry->id\" data-site-id=\"$targetEntry->siteId\">", $translatedValue);
-                }
-            }
-        }
-
-        return $translatedValue;
-    }
-
-    private function translateContentBlock(Element $source, \craft\fields\ContentBlock $field, Site $sourceSite, Site $targetSite)
-    {
-        $serialized = $source->getSerializedFieldValues([$field->handle])[$field->handle];
-        $fields = [];
-        $this->processFieldLayoutFields($source->getFieldValue($field->handle), $fields, $sourceSite, $targetSite);
-
-        foreach ($fields as $fieldHandle => $value) {
-            $serialized['fields'][$fieldHandle] = $value;
-        }
-
-        return $serialized;
     }
 
     public function onBeforeElementTranslation(Element $source, Site $sourceSite, Site $targetSite, bool $isRootElement): bool
@@ -664,5 +449,33 @@ class TranslateService extends Component
     protected function getProviderSettings(): \digitalpulsebe\craftmultitranslator\records\ProviderSettings
     {
         return MultiTranslator::getInstance()->settingsService->getProviderSettings();
+    }
+
+    private function registerSerializerClasses(): void
+    {
+        foreach (static::$matrixFields as $fieldClass) {
+            $this->serializers[$fieldClass] = MatrixSerializer::class;
+        }
+
+        $this->serializers['craft\fields\PlainText'] = TextSerializer::class;
+        $this->serializers['craft\redactor\Field'] = RichTextSerializer::class;
+        $this->serializers['craft\ckeditor\Field'] = RichTextSerializer::class;
+        $this->serializers['abmat\tinymce\Field'] = RichTextSerializer::class;
+        $this->serializers['verbb\hyper\fields\HyperField'] = HyperSerializer::class;
+        $this->serializers['craft\fields\Table'] = TableSerializer::class;
+        $this->serializers['craft\fields\ContentBlock'] = ContentBlockSerializer::class;
+        $this->serializers['presseddigital\linkit\fields\LinkitField'] = LinkitSerializer::class;
+        $this->serializers['ether\seo\fields\SeoField'] = EtherSeoSerializer::class;
+        $this->serializers['nystudio107\seomatic\fields\SeoSettings'] = SeomaticSerializer::class;
+        $this->serializers['verbb\vizy\fields\VizyField'] = VizySerializer::class;
+
+        // trigger event to add custom serializers
+        $event = new RegisterSerializersEvent([
+            'serializers' => $this->serializers,
+        ]);
+
+        $this->trigger(self::EVENT_REGISTER_SERIALIZERS, $event);
+
+        $this->serializers = $event->serializers;
     }
 }